+++ mirrord-layer/src/lib.rs
@@ -469,7 +469,8 @@ async fn start_layer_thread(
                     }
                 }
             } else {
-                graceful_exit!("unexpected response - expected env vars response {msg:?}");
+                let raw_issue = format!("Expected env vars response, but got {msg:?}");
+                graceful_exit!("{}{FAIL_STILL_STUCK}{}", FAIL_UNEXPECTED_RESPONSE, raw_issue);
             }
           },
           _ = sleep(Duration::from_secs(config.agent.communication_timeout.unwrap_or(30).into())) => {
@@ -558,6 +559,24 @@ pub(crate) unsafe extern "C" fn close_detour(fd: c_int) -> c_int {
 pub(crate) unsafe extern "C" fn close_nocancel_detour(fd: c_int) -> c_int {
     close_detour(fd)
 }
+
+pub(crate) const FAIL_STILL_STUCK: &str = r#"
+- If you're still stuck and everything looks fine:
+
+>> Please open a new bug report at https://github.com/metalbear-co/mirrord/issues/new/choose
+
+>> Or join our discord https://discord.com/invite/J5YSrStDKD and request help in #mirrord-help.
+
+"#;
+
+const FAIL_UNEXPECTED_RESPONSE: &str = r#"
+mirrord-layer received an unexpected response from the agent pod!
+
+- Suggestions:
+
+>> When trying to run a program with arguments in the form of `app -arg value`, run it as
+   `app -- -arg value` instead.
+"#;
 
 #[cfg(test)]
 mod tests {